небольшие изменения

+++ TestMapBox/Models/Coordinate.cs
@@ -5,9 +5,9 @@ namespace TestMapBox.Models
     public class Coordinate
     {
         [JsonProperty("lng")]
-        public double lng { get; set; }
+        public decimal lng { get; set; }
 
         [JsonProperty("lat")]
-        public double lat { get; set; }
+        public decimal lat { get; set; }
     }
 }
