+++ TestMapBox/SolutionClasses/Customer.cs
@@ -4,18 +4,19 @@
     {
         public int CustomerId { get; set; }
         public int Demand { get; set; }
-        public double X { get; set; }
-        public double Y { get; set; }
+        public decimal X { get; set; }
+        public decimal Y { get; set; }
+
         public bool IsRouted;
 
-        public Customer(double x, double y) //for depot
+        public Customer(decimal x, decimal y) //for depot
         {
             CustomerId = 0;
             X = x;
             Y = y;
         }
 
-        public Customer(int id, double x, double y, int demand)
+        public Customer(int id, decimal x, decimal y, int demand)
         {
             CustomerId = id;
             X = x;
